backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-community
langchain-pinecone
langchain-huggingface
langchain-text-splitters
langgraph
langchain-tavily
pinecone
youtube-transcript-api
